fix: id type을 Long

+++ src/main/java/project/teaming/invite/dto/AcceptInviteRequestDto.java
@@ -1,7 +1,7 @@
 package project.teaming.invite.dto;
 
 public record AcceptInviteRequestDto (
-        Integer inviteId
+        Long inviteId
 ) {
 
 }
